requirements.txt
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Machine learning
xgboost>=1.5.0
torch>=1.9.0
transformers>=4.10.0

# Bioinformatics
biopython>=1.78

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
tqdm>=4.62.0
joblib>=1.0.0

# Optional: GPU support (uncomment if using CUDA)
# torch>=1.9.0+cu111 -f https://download.pytorch.org/whl/torch_stable.html
